go-ooo - db restructure for curated DEX pairs for AdHoc queries

+++ go-ooo/database/migrations.go
@@ -1,5 +1,7 @@
 package database
 
+import "go-ooo/database/models"
+
 /*
   Migrations
 */
@@ -9,7 +11,7 @@ package database
 func (d *DB) MigrateV0ToV1() {
 	dbVers, _ := d.getCurrentDbSchemaVersion()
 	if dbVers.CurrentVersion == 0 {
-		d.V0ToV1DeleteAdhocTokenData()
+		d.DeleteAdhocTokenData()
 		_ = d.setDbSchemaVersion(1)
 	}
 }
@@ -17,20 +19,39 @@ func (d *DB) MigrateV0ToV1() {
 func (d *DB) MigrateV1ToV2() {
 	dbVers, _ := d.getCurrentDbSchemaVersion()
 	if dbVers.CurrentVersion == 1 {
-		d.V1ToV2DeleteAdhocTokenData()
+		d.DeleteAdhocTokenData()
 		_ = d.setDbSchemaVersion(2)
 	}
 }
 
-// V0ToV1DeleteAdhocTokenData is used when migrating from Db v0 to v1
-func (d *DB) V0ToV1DeleteAdhocTokenData() {
-	d.Exec("DELETE FROM dex_pairs")
-	d.Exec("DELETE FROM dex_tokens")
-	d.Exec("DELETE FROM token_contracts")
+func (d *DB) MigrateV2ToV3() {
+	dbVers, _ := d.getCurrentDbSchemaVersion()
+	if dbVers.CurrentVersion == 2 {
+		d.DeleteAdhocTokenData()
+
+		if d.Migrator().HasTable("dex_tokens") {
+			d.Migrator().DropTable("dex_tokens")
+		}
+
+		d.Migrator().DropColumn(&models.DexPairs{}, "t0_dex_token_id")
+		d.Migrator().DropColumn(&models.DexPairs{}, "t1_dex_token_id")
+		d.Migrator().DropColumn(&models.DexPairs{}, "dex_name")
+
+		_ = d.setDbSchemaVersion(3)
+	}
 }
 
-// V1ToV2DeleteAdhocTokenData is used when migrating from Db v1 to v2
-// does the same as v0 to v1 migration, because module names have changed
-func (d *DB) V1ToV2DeleteAdhocTokenData() {
-	d.V0ToV1DeleteAdhocTokenData()
+// DeleteAdhocTokenData is used when migrating from Db v0 to v1
+func (d *DB) DeleteAdhocTokenData() {
+	if d.Migrator().HasTable("dex_pairs") {
+		d.Exec("DELETE FROM dex_pairs")
+	}
+
+	if d.Migrator().HasTable("dex_tokens") {
+		d.Exec("DELETE FROM dex_tokens")
+	}
+
+	if d.Migrator().HasTable("token_contracts") {
+		d.Exec("DELETE FROM token_contracts")
+	}
 }
